Add InvalidPrefixError
Part of VRVDR-53854. Add a test message for an invalid prefix that can be used in configd unit tests.  'setMsg' is a bit of a hack, but then the message is really intended for internal use if some application provides an incorrect prefix when calling cfgcli so it serves its purpose.

(cherry picked from commit 2b520251f72d9b70894e7b9b53c769cc4eeb34e0)